Version 4.2
26.01.2022 Version 4.2 - No More Room In Hell Server + Guides
Added No More Room In Hell server and Guides
Fixed an issue where an installed server would appear as not installed on the main server panel after selecting a server installation directory.
Fixed an issue where the tool would sometimes stay open in the background after you've closed it.
Minor bug fixes.
Minor UI modifications.
Added a message to inform the user to type "quit" on steamCMD after the installation of a steam server has ended.
SvenDS can only be started with a batch file(If it doesn't exist it is created automatically with default settings).